restructure projects
