+++ benchmarks/HostedService/HostedService.csproj
@@ -14,7 +14,6 @@
       <ProjectReference Include="..\..\src\Proto.Cluster.Consul\Proto.Cluster.Consul.csproj" />
       <ProjectReference Include="..\..\src\Proto.Cluster.Identity.MongoDb\Proto.Cluster.Identity.MongoDb.csproj" />
       <ProjectReference Include="..\..\src\Proto.Cluster\Proto.Cluster.csproj" />
-      <ProjectReference Include="..\..\src\Proto.Remote.GrpcCore\Proto.Remote.GrpcCore.csproj" />
     </ItemGroup>
 
 </Project>
