+++ benchmarks/KubernetesDiagnostics/KubernetesDiagnostics.csproj
@@ -15,7 +15,6 @@
       <PackageReference Include="Proto.Cluster.Kubernetes" Version="0.26.0" />
         <PackageReference Include="Proto.Cluster.Consul" Version="0.26.0" />
       <PackageReference Include="Proto.Remote.GrpcNet" Version="0.26.0" />
-        <PackageReference Include="Proto.Remote.GrpcCore" Version="0.26.0" />
       <PackageReference Include="StackExchange.Redis" Version="2.2.50" />
     </ItemGroup>
