+++ benchmarks/RemoteBenchmark/Node2/Program.cs
@@ -12,7 +12,6 @@ using Messages;
 using Microsoft.Extensions.Logging;
 using Proto;
 using Proto.Remote;
-using Proto.Remote.GrpcCore;
 using Proto.Remote.GrpcNet;
 using ProtosReflection = Messages.ProtosReflection;
