+++ examples/ClusterGrainHelloWorld/Node1/Node1.csproj
@@ -8,7 +8,6 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\src\Proto.Cluster\Proto.Cluster.csproj" />
-    <ProjectReference Include="..\..\..\src\Proto.Remote.GrpcCore\Proto.Remote.GrpcCore.csproj" />
     <ProjectReference Include="..\..\..\src\Proto.Remote.GrpcNet\Proto.Remote.GrpcNet.csproj" />
     <ProjectReference Include="..\Messages\Messages.csproj" />
   </ItemGroup>
