+++ examples/EventStreamTopicsCluster/EventStreamTopicsCluster.csproj
@@ -9,7 +9,7 @@
     <ItemGroup>
       <ProjectReference Include="..\..\src\Proto.Cluster.Consul\Proto.Cluster.Consul.csproj" />
       <ProjectReference Include="..\..\src\Proto.Cluster\Proto.Cluster.csproj" />
-      <ProjectReference Include="..\..\src\Proto.Remote.GrpcCore\Proto.Remote.GrpcCore.csproj" />
+      <ProjectReference Include="..\..\src\Proto.Remote.GrpcCore\Proto.Remote.csproj" />
     </ItemGroup>
 
     <ItemGroup>
