+++ src/Proto.Cluster.Identity.MongoDb/Proto.Cluster.Identity.MongoDb.csproj
@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <LangVersion>9</LangVersion>
-        <TargetFrameworks>net5.0;net6.0;netstandard2.1</TargetFrameworks>
+        <TargetFrameworks>net5.0;net6.0;netcoreapp3.1</TargetFrameworks>
         <Nullable>enable</Nullable>
     </PropertyGroup>
