+++ src/Proto.Cluster.TestProvider/Proto.Cluster.TestProvider.csproj
@@ -2,7 +2,7 @@
   <PropertyGroup>
     <RootNamespace>Proto.Cluster.Consul</RootNamespace>
     <LangVersion>9</LangVersion>
-    <TargetFrameworks>net5.0;net6.0;netstandard2.1</TargetFrameworks>
+    <TargetFrameworks>net5.0;net6.0;netcoreapp3.1</TargetFrameworks>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Options" Version="6.0.0" />
