+++ src/Proto.Remote/GrpcNet/RemoteHostedService.cs
