+++ src/Proto.Remote/Proto.Remote.csproj
@@ -3,12 +3,15 @@
     <IsPackable>true</IsPackable>
     <Nullable>enable</Nullable>
     <LangVersion>9</LangVersion>
-    <TargetFrameworks>net5.0;net6.0;netstandard2.1</TargetFrameworks>
+    <TargetFrameworks>net5.0;net6.0;netcoreapp3.1</TargetFrameworks>
   </PropertyGroup>
+  
   <ItemGroup>
-    <PackageReference Include="Grpc.Core" Version="2.43.0" />
+    <PackageReference Include="Grpc.AspNetCore" Version="2.43.0" />
+    <PackageReference Include="Grpc.Core.Api" Version="2.43.0" />
     <PackageReference Include="Grpc.HealthCheck" Version="2.43.0" />
-    <PackageReference Include="Grpc.Tools" Version="2.40.0" PrivateAssets="All" />
+    <PackageReference Include="Grpc.Net.Common" Version="2.43.0" />
+    <PackageReference Include="Grpc.Tools" Version="2.43.0" PrivateAssets="All" />
     <PackageReference Include="Roslynator.Analyzers" Version="4.0.2">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
